fix(admin): remove unsupported topic sensitivity handling
Remove sensitive configuration metadata and presentation because Kafka topic configurations cannot be sensitive.
