+++ kaskade/unicodes.py
@@ -9,4 +9,3 @@ DOWN_TRIANGLE = "\u25bc"
 DIAMOND = "\u25c8"
 PIPE = "│"
 WARNING = "\u26a0"
-LOCK = "\U0001f512"
